Revert 2b4241c..c1bfd5d
This rolls back to commit 2b4241c347c643d5586527271b79e5d2c2e2d84b.
