+++ src/main/java/org/apache/mesos/hdfs/executor/NodeExecutor.java
@@ -9,14 +9,7 @@ import org.apache.mesos.ExecutorDriver;
 import org.apache.mesos.MesosExecutorDriver;
 import org.apache.mesos.Protos.*;
 import org.apache.mesos.hdfs.config.SchedulerConf;
-import org.apache.mesos.hdfs.state.LiveState;
 import org.apache.mesos.hdfs.util.HDFSConstants;
-
-import java.io.IOException;
-import java.net.Socket;
-import java.util.Set;
-import java.util.Timer;
-import java.util.TimerTask;
 
 /**
  * The executor for a Basic Node (either a Journal Node or Data Node).
@@ -25,6 +18,7 @@ import java.util.TimerTask;
 public class NodeExecutor extends AbstractNodeExecutor {
   public static final Log log = LogFactory.getLog(NodeExecutor.class);
   private Task task;
+
   private Task dataNodeTask;
 
   /**
@@ -32,8 +26,8 @@ public class NodeExecutor extends AbstractNodeExecutor {
    * 
    **/
   @Inject
-  NodeExecutor(SchedulerConf schedulerConf, LiveState liveState) {
-    super(schedulerConf, liveState);
+  NodeExecutor(SchedulerConf schedulerConf) {
+    super(schedulerConf);
   }
 
   /**
@@ -62,7 +56,7 @@ public class NodeExecutor extends AbstractNodeExecutor {
           .setState(TaskState.TASK_RUNNING)
           .setData(taskInfo.getData()).build());
     } else {
-      // for other tasks on this node, just start them normally
+      //for other tasks on this node, just start them normally
       driver.sendStatusUpdate(TaskStatus.newBuilder()
           .setTaskId(taskInfo.getTaskId())
           .setState(TaskState.TASK_RUNNING)
@@ -86,45 +80,8 @@ public class NodeExecutor extends AbstractNodeExecutor {
     String messageStr = new String(msg);
     // launch tasks on init (primary NN), launch tasks on run (secondary NN)
     if (messageStr.equals(HDFSConstants.DATA_NODE_INIT_MESSAGE)) {
-      // start the data node after name nodes are visible
-      Timer timer = new Timer(true);
-      TimerTask waitForNameNodes = new NNPortCheckTask(
-          driver,
-          liveState.getNameNodeDomainNames(),
-          HDFSConstants.NAME_NODE_HTTP_PORT);
-      timer.scheduleAtFixedRate(waitForNameNodes, 0, 15000);
-    }
-  }
-
-  private class NNPortCheckTask extends TimerTask {
-    ExecutorDriver driver;
-    Set<String> hosts;
-    int port;
-
-    public NNPortCheckTask(ExecutorDriver driver, Set<String> hosts, int port) {
-      this.driver = driver;
-      this.hosts = hosts;
-      this.port = port;
-    }
-
-    @Override
-    public void run() {
-      boolean success = true;
-      for (String host : hosts) {
-        log.info("Checking for " + host + " at port " + port);
-        try (Socket connected = new Socket(host, port)) {
-          log.info("Successfully found " + host + " at port " + port);
-        } catch (SecurityException | IOException e) {
-          log.info("Couldn't resolve host " + host + " at port " + port);
-          success = false;
-          break;
-        }
-      }
-      if (success) {
-        log.info("Successfully reached name nodes. Starting process: " + dataNodeTask.taskInfo.getName());
-        startProcess(driver, dataNodeTask);
-        this.cancel();
-      }
+      // start the data node
+      startProcess(driver, dataNodeTask);
     }
   }
 }
